[powdr] Powdr labs/resolve conflicts 27 08

+++ crates/zkvm/entrypoint/src/syscalls/fptower.rs
@@ -6,7 +6,7 @@ use core::arch::asm;
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bls12381_fp_addmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bls12381_fp_addmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -26,7 +26,7 @@ pub extern "C" fn syscall_bls12381_fp_addmod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bls12381_fp_submod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bls12381_fp_submod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -46,7 +46,7 @@ pub extern "C" fn syscall_bls12381_fp_submod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bls12381_fp_mulmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bls12381_fp_mulmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -66,7 +66,7 @@ pub extern "C" fn syscall_bls12381_fp_mulmod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bls12381_fp2_addmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bls12381_fp2_addmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -86,7 +86,7 @@ pub extern "C" fn syscall_bls12381_fp2_addmod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bls12381_fp2_submod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bls12381_fp2_submod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -106,7 +106,7 @@ pub extern "C" fn syscall_bls12381_fp2_submod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bls12381_fp2_mulmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bls12381_fp2_mulmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -126,7 +126,7 @@ pub extern "C" fn syscall_bls12381_fp2_mulmod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bn254_fp_addmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bn254_fp_addmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -146,7 +146,7 @@ pub extern "C" fn syscall_bn254_fp_addmod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bn254_fp_submod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bn254_fp_submod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -166,7 +166,7 @@ pub extern "C" fn syscall_bn254_fp_submod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bn254_fp_mulmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bn254_fp_mulmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -186,7 +186,7 @@ pub extern "C" fn syscall_bn254_fp_mulmod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bn254_fp2_addmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bn254_fp2_addmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -206,7 +206,7 @@ pub extern "C" fn syscall_bn254_fp2_addmod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bn254_fp2_submod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bn254_fp2_submod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -226,7 +226,7 @@ pub extern "C" fn syscall_bn254_fp2_submod(x: *mut u32, y: *const u32) {
 /// The result is written over the first input.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_bn254_fp2_mulmod(x: *mut u32, y: *const u32) {
+pub extern "C" fn syscall_bn254_fp2_mulmod(x: *mut u64, y: *const u64) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
